added new entity "art"; improved movement system for AI to better match how player moves; added new entity definition; added player velocity to debug gui.

+++ core/src/xyz/vec3d/game/model/combat/CombatSystem.java
@@ -98,7 +98,7 @@ public class CombatSystem implements IMessageReceiver {
         }
         //If an entity is being hit, get its bonuses
         if (entityBeingHit instanceof Enemy) {
-            targetBaseDefenseBonuses = (float[]) DefinitionLoader.getEntityDefinition(((Enemy) entityBeingHit).getId()).getProperty(DefinitionProperty.BONUSES);
+            targetBaseDefenseBonuses = (int[]) DefinitionLoader.getEntityDefinition(((Enemy) entityBeingHit).getId()).getProperty(DefinitionProperty.BONUSES);
         }
 
         switch (type) {
